start drv func

+++ drv/struct.hpp
@@ -1,29 +1,30 @@
 #pragma once
 #include <cstdint>
+#include "Windows.h"
 
 struct Struct10317 {
-	uintptr_t Member1031A;
-	uintptr_t Member1031B;
-	uintptr_t Member1031C;
-	uintptr_t Member1031D;
+	DWORD Member1031A;
+	DWORD Member1031B;
+	DWORD Member1031C;
+	DWORD Member1031D;
 };
 
-struct Struct10318 {
-	int Member1031E;
-	int Member1031F;
-	int Member10320;
-	int Member10321;
-	int Member10322;
-	int Member10323;
-	int Member10324;
+struct Struct10318 {	// _SERVICE_STATUS
+	DWORD Member1031E;	// dwServiceType
+	DWORD Member1031F;	// dwCurrentState
+	DWORD Member10320;	// dwControlsAccepted
+	DWORD Member10321;	// dwWin32ExitCode
+	DWORD Member10322;	// dwServiceSpecificExitCode
+	DWORD Member10323;	// dwCheckPoint
+	DWORD Member10324;	// dwWaitHint
 };
 
 struct Struct10319 {
-	int Member10325;
-	int Member10326;
-	int Member10327;
-	int Member10328;
-	int Member10329;
-	int Member1032A;
-	int Member1032B;
+	DWORD Member10325;
+	DWORD Member10326;
+	DWORD Member10327;
+	DWORD Member10328;
+	DWORD Member10329;
+	DWORD Member1032A;
+	DWORD Member1032B;
 };
